mark IMuninNodeBuilder obsolete
